feat : 구간에 등록된 역 삭제 시 오류 발생

+++ src/main/java/nextstep/subway/applicaion/StationService.java
@@ -3,19 +3,21 @@ package nextstep.subway.applicaion;
 import lombok.RequiredArgsConstructor;
 import nextstep.subway.applicaion.dto.StationRequest;
 import nextstep.subway.applicaion.dto.StationResponse;
+import nextstep.subway.domain.section.SectionRepository;
 import nextstep.subway.domain.station.Station;
 import nextstep.subway.domain.station.StationRepository;
-import nextstep.subway.utils.ObjectMapUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
 @Transactional(readOnly = true)
 public class StationService {
     private final StationRepository stationRepository;
+    private final SectionRepository sectionRepository;
 
     @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
@@ -23,11 +25,14 @@ public class StationService {
     }
 
     public List<StationResponse> findAllStations() {
-        return ObjectMapUtils.mapLooseAll(stationRepository.findAll(), StationResponse.class);
+        return stationRepository.findAll().stream().map(StationResponse::new).collect(Collectors.toList());
     }
 
     @Transactional
     public void deleteStationById(Long id) {
+        if (sectionRepository.existsByUpStation_Id(id) || sectionRepository.existsByDownStation_Id(id)) {
+            throw new IllegalArgumentException("station.used.section");
+        }
         stationRepository.deleteById(id);
     }
